trace: allow settings "span.name" tag to increase interoperability (#15)
This change permits changing the span's name when it is exported to Datadog using the "span.name" tag. This is useful in cases when multiple exporters are used because Datadog requires consistent top-level naming for spans, while other systems (like Stackdriver) expect what we call resource for span name. With this change, the resource can safely be used for the span name (as most systems expect it) while also adding the "span.name" tag specifically for Datadog.